apps/word_buddy/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-dotenv
pydantic
google-cloud-firestore
python-multipart
pyOpenSSL
nltk
